Remove unnecessary variable pr and use cons for printf&panic
