+++ kernel/console.h
@@ -1,0 +1,15 @@
+#include "types.h"
+#include "spinlock.h"
+
+struct console {
+  struct spinlock lock;
+  
+  // input
+#define INPUT_BUF 128
+  char buf[INPUT_BUF];
+  uint r;  // Read index
+  uint w;  // Write index
+  uint e;  // Edit index
+
+  int locking;
+};
