+++ kernel/main.c
@@ -12,7 +12,6 @@ main()
 {
   if(cpuid() == 0){
     consoleinit();
-    printfinit();
     printf("\n");
     printf("xv6 kernel is booting\n");
     printf("\n");
